refactor(inference): hoist shared thermal_sampler into bench::run (PR #208 review)
The QNN and llama.cpp benchmark examples each carried a verbatim copy of
the background thermal-sampler loop (tokio interval + oneshot stop +
sysfs reads). Extract it once as `bench::run::spawn_thermal_sampler`,
which returns a stop sender + join handle, and have both examples call
it. Behaviour is identical; the helper additionally breaks cleanly when
the sender is dropped. Prunes the now-unused thermal/oneshot/Path
imports from both examples.

Addresses the sole actionable finding from the PR #208 review. The other
two review notes were assessed as non-defects: the multi-chunk timing
test floors against `tokio::time::sleep`'s guaranteed minimum (so it
cannot flake), and the 15-min default duration mirrors qnn_bench by
design.
