+++ src/crates/primer-inference/examples/qnn_bench.rs
@@ -34,12 +34,10 @@ use primer_core::inference::{GenerationParams, InferenceBackend};
 use primer_inference::QnnBackend;
 use primer_inference::bench::{
     BENCH_MAX_TOKENS, BenchReport, BenchTargets, DEFAULT_BENCH_SYSTEM_PROMPT,
-    DEFAULT_DURATION_SECS, DEFAULT_PROMPTS_PATH, PromptMeasurement, THERMAL_SAMPLE_INTERVAL,
-    THERMAL_SYSFS_DIR, ThermalSample, evaluate, format_report, load_bench_prompts, measure_prompt,
-    peak_temp_celsius, read_thermal_zones, thermal_csv,
+    DEFAULT_DURATION_SECS, DEFAULT_PROMPTS_PATH, PromptMeasurement, evaluate, format_report,
+    load_bench_prompts, measure_prompt, peak_temp_celsius, spawn_thermal_sampler, thermal_csv,
 };
 use primer_inference::qnn::bench::qnn_targets;
-use tokio::sync::oneshot;
 
 /// Conventional QAIRT lib subdirectory under the bundle's parent, matching
 /// the AI Hub apps asset layout (mirrors the CLI/GUI default).
@@ -151,8 +149,7 @@ async fn run() -> Result<bool, Box<dyn std::error::Error>> {
     // Start the thermal sampler. It owns its own clock and returns the
     // collected samples when signalled to stop.
     let started = Instant::now();
-    let (stop_tx, stop_rx) = oneshot::channel::<()>();
-    let sampler = tokio::spawn(thermal_sampler(started, stop_rx));
+    let (stop_tx, sampler) = spawn_thermal_sampler(started);
 
     let params = GenerationParams {
         max_tokens: BENCH_MAX_TOKENS,
@@ -215,22 +212,4 @@ async fn run() -> Result<bool, Box<dyn std::error::Error>> {
         format_report("QNN benchmark", &report, &targets, &verdict)
     );
     Ok(verdict.all_pass())
-}
-
-/// Background thermal sampler. Ticks every [`THERMAL_SAMPLE_INTERVAL`],
-/// reading every `thermal_zone*/temp` under [`THERMAL_SYSFS_DIR`], until
-/// the stop signal fires. Returns all collected samples.
-async fn thermal_sampler(started: Instant, mut stop: oneshot::Receiver<()>) -> Vec<ThermalSample> {
-    let mut samples = Vec::new();
-    let mut ticker = tokio::time::interval(THERMAL_SAMPLE_INTERVAL);
-    loop {
-        tokio::select! {
-            _ = ticker.tick() => {
-                let elapsed = started.elapsed().as_secs_f64();
-                samples.extend(read_thermal_zones(Path::new(THERMAL_SYSFS_DIR), elapsed));
-            }
-            _ = &mut stop => break,
-        }
-    }
-    samples
 }
